# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  clap_internal_modules_playback_dto.ScheduleSongRequest:
    properties:
      duration_ms:
        description: |-
          DurationMs is the expected playback window in milliseconds.
          Providing it enables overlap detection against other schedules.
          Zero means "unknown duration" and overlap check is skipped.
        type: integer
      match_id:
        type: string
      scheduled_at:
        type: string
      song_id:
        type: string
    required:
    - match_id
    - scheduled_at
    - song_id
    type: object
...
  clap_internal_shared_response.ErrorDetail:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  clap_internal_shared_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/clap_internal_shared_response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Clap API Support
  description: Clap match-day experience backend (auth, clubs, matches, realtime,
    shop).
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Clap Backend API
  version: "1.0"
paths:
...
  /api/v1/media/{id}/playback-url:
    get:
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get playback URL
      tags:
      - media
  /api/v1/media/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Media file
        in: formData
        name: file
        required: true
        type: file
      - description: Media type
        in: formData
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Upload media
      tags:
      - media
  /api/v1/mobile/home/club:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Club home
      tags:
      - mobile-home
  /api/v1/mobile/home/stadium:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Stadium home
      tags:
      - mobile-home
...
  /api/v1/realtime/admin/cleanup/all:
    post:
      description: Admin only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Cleanup all retention data
      tags:
      - realtime
  /api/v1/realtime/admin/cleanup/heartbeats:
    post:
      description: Admin only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Cleanup heartbeats
      tags:
      - realtime
  /api/v1/realtime/admin/cleanup/realtime-events:
    post:
      description: Admin only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Cleanup realtime events
      tags:
      - realtime
  /api/v1/realtime/admin/cleanup/scheduler-events:
    post:
      description: Admin only
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Cleanup scheduler events
      tags:
      - realtime
...
  /api/v1/realtime/session/{matchId}:
    get:
      parameters:
      - description: Match ID
        in: path
        name: matchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Realtime session recovery
      tags:
      - realtime
...
  /api/v1/songs/{id}/audio:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: string
      - description: Audio file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Upload song audio
      tags:
      - media
...
  /api/v1/songs/schedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_playback_dto.ScheduleSongRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Schedule song playback
      tags:
      - playback
  /api/v1/songs/schedule/{id}:
    delete:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Cancel scheduled song
      tags:
      - playback
  /api/v1/songs/schedule/upcoming:
    get:
      parameters:
      - description: Match ID
        in: query
        name: match_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Upcoming scheduled songs
      tags:
      - playback
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"